test cases added for genres

+++ routes/genres.js
@@ -2,6 +2,7 @@ import { Router } from 'express';
 import validate, { Genre } from '../models/genre.js';
 import { auth } from '../middleware/auth.js';
 import { admin } from '../middleware/admin.js';
+import mongoose from 'mongoose';
 
 
 const router = Router();
@@ -11,18 +12,18 @@ router.get('/', async (req, res) => {
   res.send(genres);
 });
 
-router.post('/',auth, async (req, res) => {
-  const { error } = validate(req.body); 
+router.post('/', auth, async (req, res) => {
+  const { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   let genre = new Genre({ name: req.body.name });
   genre = await genre.save();
-  
+
   res.send(genre);
 });
 
 router.put('/:id', async (req, res) => {
-  const { error } = validate(req.body); 
+  const { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   const genre = await Genre.findByIdAndUpdate(req.params.id, { name: req.body.name }, {
@@ -30,11 +31,14 @@ router.put('/:id', async (req, res) => {
   });
 
   if (!genre) return res.status(404).send('The genre with the given ID was not found.');
-  
+
   res.send(genre);
 });
 
 router.delete('/:id', [auth, admin], async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id))
+    return res.status(400).send('Invalid Id.');
+
   const genre = await Genre.findByIdAndRemove(req.params.id);
 
   if (!genre) return res.status(404).send('The genre with the given ID was not found.');
